requirements.txt
fastapi
uvicorn
sqlalchemy
langchain
langchain-community
langchain-groq
python-dotenv
